Add new test file for exceptions in global scope
